tools/photo_pipeline/requirements.txt
Pillow>=11.0.0
rembg>=2.0.59
onnxruntime>=1.22.0
